extract the main() in a separate file
so that top-level seminator code (in seminator.cpp) is
compiled in the library
